pyproject.toml
[project]
name = "freecad-mcp"
version = "0.1.16"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "neka-nat", email = "nekanat.stock@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.2",
    "pillow>=10.4.0",
    "validators>=0.34.0",
]

[project.scripts]
freecad-mcp = "freecad_mcp.server:main"

[tool.hatch.build.targets.sdist]
exclude = ["assets", "results"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
